test_priceindex/pyproject.toml
[tool.poetry]
name = "testpriceindex"
version = "0.1.0"
description = ""
authors = ["Henrik Anfinsen <henrik@solgt.no>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9.0,<3.11"
flask = "^2.2.3"
geopandas = "^0.12.2"
matplotlib = "^3.7.1"
folium = "0.14.0"
"solgt.db" = {path = "../../py/lib/db", develop = true}
"solgt.priceindex" = {path = "../../py/lib/priceindex", develop = true}
"solgt.timeseries" = {path = "../../py/lib/timeseries", develop = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
